refactor(neon_notifications): Refactor package structure
