+++ packages/neon/neon_notifications/lib/src/blocs/notifications.dart
@@ -1,4 +1,11 @@
-part of '../neon_notifications.dart';
+import 'dart:async';
+
+import 'package:neon/blocs.dart';
+import 'package:neon/models.dart';
+import 'package:neon/utils.dart';
+import 'package:neon_notifications/src/options.dart';
+import 'package:nextcloud/notifications.dart' as notifications;
+import 'package:rxdart/rxdart.dart';
 
 abstract interface class NotificationsBlocEvents {
   void deleteNotification(final int id);
