+++ packages/neon/neon_notifications/lib/src/options.dart
@@ -1,4 +1,5 @@
-part of 'neon_notifications.dart';
+import 'package:neon/models.dart';
+import 'package:neon/settings.dart';
 
 class NotificationsAppSpecificOptions extends NextcloudAppOptions implements NotificationsOptionsInterface {
   NotificationsAppSpecificOptions(super.storage) {
